Finished project
A very simple game with very simple code. Would have liked to incorporate more uses of the music note files and mini games (like having the player read a long piece of sheet music that would scroll off-screen as the player hits the right notes, and using a metronome/timer to account for rhythm and tempo).